rustdoc: collect trait impls as an early pass
